V1.16!
Honestly, v1.08 and v1.16 could have been wrapped up in the same update, but whatever.
Additions:
REGI: Defines a register of the ID X. Takes one arguement
SETR: Sets the value of register X1 to X2. Takes two arguments.
RVAL: MultiOp, returns the value of register X. Takes one argument.
Added ERR 25, 26 & 27.
Changes:
ERR 12 is now only for RULES, and ERR 21 is now the ID already in use error for cmpxs. This is because the previous ERR 21 was untriggerable.