v1.30!
A rather small update, 1.30 does bring some nice features to the table.
ROOT, the newest multiOp, is also the first math multiOp since AE1 (Oct. 2). It takes two arguments, the first is the value to be rooted, and the second is the root amount. i.e. ROOT 4i 2i is the square root of 4, ROOT 125i 3i is the cube root of 125, etc.
The second added feature is the addition of STOP, which pauses execution until the user presses a key. This functionality was removed from EXIT.

Additional changes include:
ERR 26 has been standardized within the interpreter
Various bugs have been patched
towcbl-mode now supports the added features
math library now must be specified when compiling (-lm for gcc)